Add logic to the calculator to make it usable
This will allow the calculator to make use of
addition, subtraction, multiplication, and
division.

However, the calculator currently has some bugs
in this state. The calculator cannot perform some
special actions such as, chaining multiple
operators in a single calculation, and tapping the
'=' button to make consecutive calculations with
the same operator and current result.

I shall work towards fixing these bugs and adding
more features in the next commit.