Add new feature to equals button and space to code
If the user selects a number, followed by an
operator, and then presses the equals button, the
calculator will use the current number in the
display as the second number for the calculation.

This commit will also space out some of the lines
in the code, primarily in the functions and event
listeners, to make the code more readable.