Add keyboard support and modify display font size
The user can now input numbers and operators using
their keyboard.

The user can now only enter up to 14 digit numbers
(100 trillion), and the font size of the display
will dynamically change as the size of the number
or result changes. This will prevent the number in
the display from overflowing outside of the
calculator.